% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimage.R
\name{as.POSIXct.Pimage}
\alias{as.POSIXct.Pimage}
\title{as.POSIXct for Pimage}
\usage{
\method{as.POSIXct}{Pimage}(x, ...)
}
\arguments{
\item{x}{Pimage object}

\item{...}{ignored}
}
\description{
Convert Pimage to POSIXct
}
\details{
Coercion method to extract times from Pimage object. The type is
inferred from the internal storage.
}
\seealso{
\code{estelleMetropolis} and \code{Pimage}
}
