% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{chainApply}
\alias{chainApply}
\alias{chainApply2}
\title{Apply a function to chains}
\usage{
chainApply(s1, f, collapse = TRUE, thin = 1, chains = NULL)

chainApply2(s1, s2, f, collapse = TRUE, thin = 1, chains = NULL)
}
\arguments{
\item{s1}{a single chain or a list of parallel chains generated by
\code{estelleMetropolis} or \code{stellaMetropolis}.}

\item{f}{the function to apply to samples}

\item{collapse}{whether to collapse parallel chains to a single chain.}

\item{thin}{rate at which to thin the sample.}

\item{chains}{the set of chains to retain, or \code{NULL}.}

\item{s2}{a single chain or a list of parallel chains generated by
\code{estelleMetropolis} or \code{stellaMetropolis}.}
}
\description{
Apply a function to chains of samples
}
\details{
The function \code{chainApply} applies a function of a single
argument to the samples of a chain or a list of parallel chains
generated by \code{estelleMetropolis} or
\code{stellaMetropolis}.

The function \code{chainApply2} applies a function of two
arguments to samples of two chain or two lists of parallel chains
generated by \code{estelleMetropolis} or
\code{stellaMetropolis}.
}

