% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{chainCoda}
\alias{chainCoda}
\title{Export to Coda}
\usage{
chainCoda(s)
}
\arguments{
\item{s}{a list of chains generated by \code{estelleMetropolis}
or \code{stellaMetropolis}.}
}
\value{
a \pkg{coda} object.
}
\description{
Convert to Coda objects
}
\details{
Convert samples generated by \code{estelleMetropolis} or
\code{stellaMetropolis} to a \pkg{coda} object.
}

