% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{chainSummary}
\alias{chainAcceptance}
\alias{chainBcov}
\alias{chainCollapse}
\alias{chainCov}
\alias{chainLast}
\alias{chainSummary}
\alias{chainTail}
\title{Manipulate MCMC samples}
\usage{
chainSummary(s)

chainTail(s, discard = 0, thin = 1)

chainLast(s)

chainCollapse(s, collapse = TRUE, discard = 0, thin = 1, chains = NULL)

chainCov(s, discard = 0, chains = NULL)

chainBcov(s, discard = 0, chains = NULL)

chainAcceptance(s, collapse = FALSE, chains = NULL)
}
\arguments{
\item{s}{a single chain or a list of parallel chains generated by
\code{estelleMetropolis} or \code{stellaMetropolis}.}

\item{discard}{number of initial samples to discard.}

\item{thin}{rate at which to thin the sample.}

\item{collapse}{whether to collapse parallel chains to a single chain}

\item{chains}{the set of chains to retain, or \code{NULL}.}
}
\value{
\code{chainSummary} returns a summary of the sample
\code{chainTail} discards the initial samples from each chain
\code{chainLast} returns the last sample for each location in each chain
\code{chainCollapse} collapses multiple chains into a single sample
\code{chainCov} returns the covariance of the parameters location by location as a pxpxn array.
\code{chainBcov} returns the joint covariance of the parameters as an (np)x(np) array.
\code{chainAcceptance} returns the acceptance rate in the (thinned) chain
}
\description{
Manipulate the samples generated by the Metropolis samplers.
}
\details{
These functions provide some basic operations on the samples
generated by the Metropolis samplers for Estelle and Stella.
}

