% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{coord}
\alias{coord}
\title{Geolocation Estimation by the Threshold Method}
\usage{
coord(tFirst, tSecond, type, degElevation = -6)
}
\arguments{
\item{tFirst}{factor or character vector representing the times of
the first twilight in the format "Y-m-d H:M:S".}

\item{tSecond}{factor or character vector representing the times
of the second twilight in the format "Y-m-d H:M:S".}

\item{type}{vector with elements 1 or 2, defining the elements of
\code{tFirst} as sunrise or sunset respectively.}

\item{degElevation}{sun elevation angle (90-zenith).}
}
\value{
location estimates stored as a two column (lon,lat) matrix.
}
\description{
Estimate location from two consecutive twilights by a threshold
method
}
\details{
This package and the \pkg{GeoLight} package provide some common
functionality, but are based on different astronomical
approximations.  This function provides a drop-in replacement for
the \code{coord} function from the \pkg{GeoLight} package to allow
easy comparison of the two systems.
}

