% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_codes.R
\docType{data}
\name{countries_codes}
\alias{countries_codes}
\title{Countries codes decodification}
\format{
A data frame with 136 rows and 2 columns:
\describe{
\item{country.code}{Countries codes to be used in \code{\link{mobility_var}}}
\item{country.region}{Names of the countries or regions}
}
}
\source{
\url{https://www.google.com/covid19/mobility/}
}
\usage{
countries_codes
}
\description{
A data frame containing all posibles countries codes to be used in \code{\link{mobility_var}} and their respective countries names
diamonds.
}
\keyword{datasets}
