% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{curveModel}
\alias{curveModel}
\alias{curveModel0}
\title{Curve Model Structure}
\usage{
curveModel(
  time,
  light,
  segment,
  calibration,
  alpha,
  beta,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  x0 = NULL,
  z0 = NULL,
  fixedx = FALSE,
  dt = NULL
)

curveModel0(
  time,
  light,
  segment,
  calibration,
  alpha,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  fixedx = FALSE
)
}
\arguments{
\item{time}{vector of sample times as POSIXct.}

\item{light}{vector of observed (log) light levels.}

\item{segment}{vector of integers that assign observations to
twilight segments.}

\item{calibration}{function that maps zenith angles to expected
light levels.}

\item{alpha}{parameters of the twilight model.}

\item{beta}{parameters of the behavioural model.}

\item{logp.x}{function to evaluate any additional contribution to
the log posterior from the twilight locations.}

\item{logp.z}{function to evaluate any additional contribution to
the log posterior from the intermediate locations.}

\item{x0}{suggested starting points for twilight locations.}

\item{z0}{suggested starting points for intermediate locations.}

\item{fixedx}{logical vector indicating which twilight locations
to hold fixed.}

\item{dt}{time intervals for speed calculation in hours.}
}
\value{
a list with components
\item{\code{logpx}}{function to evaluate the contributions to the
log posterior from the twilight model}
\item{\code{logpz}}{function to evaluate the contributions to the
log posterior from the prior for the z locations}
\item{\code{estelle.logpb}}{function to evaluate contribution to
the log posterior from the behavioural model for estelle.}
\item{\code{stella.logpb}}{function to evaluate contribution to
the log posterior from the behavioural model for stella.}
\item{\code{fitted}}{function to evaluate the fitted values for a
given set of locations.}
\item{\code{fixedx}}{a logical vector indicating which locations
should remain fixed.}
\item{\code{x0}}{an array of initial twilight locations.}
\item{\code{z0}}{an array of initial intermediate locations.}
\item{\code{time}}{the sample times.}
\item{\code{light}}{the recorded light levels.}
\item{\code{segment}}{vector of integers that assign observations
to twilight segments.}
}
\description{
Light Curve Model Structures for Stella and Estelle
}
\details{
Stella and Estelle require a model structure that describes the
model being fitted. These function generate basic model structures
for curve fitting methods that should provide a suitable
starting point for most analyses.

The \code{curveModel} function constructs a model structure
assuming that each twilight profile is associated with a single
location.  The errors in observed log light level are assumed to
be Normally distributed about their expected value.

The initialization locations \code{x0} and \code{z0} must be
consistent with any other constraints imposed by the data.

Both Estelle and Stella variants of the model assume that the
average speed of travel between successive locations is Gamma
distributed, and for Estelle models, the change in bearing
(degrees) along the dog-leg path segments can be assumed Normally
distributed with mean zero.  By default, the speed of travel is
calculated based on the time intervals between the twilights (in
hours), but the intervals of time actually available for travel
can be specified directly with the \code{dt} argument.

If \code{beta} is a vector, then \code{beta[1]} and \code{beta[2]}
specify the shape and rate of the Gamma distribution of speeds.
If \code{beta} has three elements, then \code{beta[3]} specifies
the standard deviation of the change in bearing (in degrees) along
dog leg paths.

The \code{curveModel0} function constructs only the non-movement
elements of the model, and can be used as a basis for those
wishing to experiment with alternative movement models.
}
