% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pimage.R
\name{cut.Pimage}
\alias{cut.Pimage}
\title{cut.POSIXt for Pimage}
\usage{
\method{cut}{Pimage}(x, breaks, ...)
}
\arguments{
\item{x}{Pimage}

\item{breaks}{an interval specification, see \code{\link{cut.POSIXt}}}

\item{...}{pass arguments to \code{\link{cut.POSIXt}}}
}
\value{
RasterLayer or RasterBrick, for multiple or single level cut respectively
}
\description{
cut.Pimage
}
\details{
Cut a Pimage object based on a date-time input breaks character string, and return a multi-layer raster
}
