% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_try.R
\name{driver_try}
\alias{driver_try}
\title{Automate driver creation and geckodriver selection}
\usage{
driver_try()
}
\value{
"Tries for river connection"
}
\description{
Automate driver creation and geckodriver selection
}
\examples{
driver_try()
}
