% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pimage.R
\name{[[.Pimage}
\alias{[[.Pimage}
\title{Extract a single Pimage time step.}
\usage{
\method{[[}{Pimage}(x, i, ..., exact = TRUE)
}
\arguments{
\item{x}{Pimage}

\item{i}{integer index}

\item{...}{ignored}

\item{exact}{ignored}
}
\value{
Pimage
}
\description{
Extract a part of a Pimage object.
}

