% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\docType{data}
\name{ElephantSeal1}
\alias{ElephantSeal1}
\alias{ElephantSeal1calib}
\alias{ElephantSeal2}
\alias{ElephantSeal2calib}
\title{Southern Elephant seal tag data}
\format{\code{ElephantSeal1} A data frame with 3 columns.  The
columns represent
\tabular{rl}{
\code{time} \tab times of measurement \cr
\code{light} \tab  (log) light values \cr
\code{segment} \tab integer indicating sequence of twilight periods \cr
}
\code{ElephantSeal2} This tag is similar to \code{ElephantSeal1}
but also has columns
\tabular{rl}{
\code{depth} \tab depth in the water column in metres (positive) \cr
\code{temp} \tab temperature in the water column in degrees celcius \cr
}
\code{ElephantSeal1calib} and \code{ElephantSealcalib2} A data
frame with 2 columns.  The columns represent
\tabular{rl}{
\code{zenith} \tab zenith values \cr
\code{light} \tab  (log) light values \cr
}}
\description{
Light data from the foraging trips of the Southern elephant seal.
}
\details{
Light intensity measurements over time from an archival tag on a
Southern elephant seal (\emph{Mirounga leonina}).  The seal was
tagged at the isthmus on Macquarie Island (158.950E, 54.5S), with
data from a time-depth-recorder (Mk9 TDR; Wildlife Computers,
Seattle, WA, USA).  These tags provides regular time series of
measurements of depth, water temperature, and ambient light
level. The original data for \code{ElephantSeal1} were processed
to remove values at depths greater than 15m and to classify
periods of twilight. The data for \code{ElephantSeal2} have also
been processed for twilight periods. The seals makes one single
foraging trip, returning to the isthmus where they were tagged.
Data recorded while the seal is at the isthmus are used for
calibration (\code{\link{ElephantSeal1calib}}).

These data supplied courtesy of Mark Hindell, Institute of Marine
and Antarctic Studies, University of
Tasmania. \code{ElephantSeal1} is B362_99 and \code{ElephantSeal2}
is C699_02
}
\keyword{data}

