% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lattice.R
\name{essie}
\alias{essie}
\title{Essie}
\usage{
essie(model, grid, epsilon1 = 0.001, epsilon2 = 1e-16,
  verbose = interactive())
}
\arguments{
\item{model}{a model structure generated by \code{essieThresholdModel}.}

\item{grid}{a raster object defining the grid}

\item{epsilon1}{likelihood tolerance}

\item{epsilon2}{transition probability tolerance.}

\item{verbose}{report progress at prompt?}
}
\value{
a list with elements
\item{\code{grid}}{a raster object that defines the grid.}
\item{\code{times}}{the times corresponding to the location estimates.}
\item{\code{lattice}}{a sparse grid representation of the posterior location probabilities.}
}
\description{
Fit an Essie model
}
\details{
Essie is a discrete analog of Stella that only considers locations
on a lattice of grid points.

The posterior probability that the tag is at a particular location
is determined by a two pass recursive algorithm.  The forward
sweep propogates location information forward in time, the
backward sweep propogates location information backward in time,
and the full posterior is a compromise of these two.

The method is only approximate and is controlled by the two
tolerances \code{epsilon1} and \code{epsilon2}.  The first limits
the precision of the likelihood - locations for which the
likelihood fall below \code{epsilon1} of its maximum are ignored.
The second limits the precision of the transition probabilities -
locations for which the probability below \code{epsilon2} of the
maximum are ignored in the movement calculation.  The smaller
these parameters are set, the slower but more accurate the fit.
}

