% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\name{essieCurveModel}
\alias{essieCurveModel}
\title{Curve Model Structures (Essie)}
\usage{
essieCurveModel(
  time,
  light,
  segment,
  calibration,
  alpha,
  beta,
  logp0 = function(k, x) 0,
  x0,
  fixed = FALSE,
  dt = NULL
)
}
\arguments{
\item{time}{vector of sample times as POSIXct.}

\item{light}{vector of observed (log) light levels.}

\item{segment}{vector of integers that assign observations to
twilight segments.}

\item{calibration}{function that maps zenith angles to expected
light levels.}

\item{alpha}{parameters of the twilight model.}

\item{beta}{parameters of the behavioural model.}

\item{logp0}{function to evaluate any additional contribution to
the log posterior from the twilight locations.}

\item{x0}{suggested starting points for twilight locations.}

\item{fixed}{logical vector indicating which twilight locations
to hold fixed.}

\item{dt}{time intervals for speed calculation in hours.}
}
\value{
a list with components
\item{\code{logpk}}{function to evaluate the contributions to the
log posterior from the k-th twilight}
\item{\code{logpbk}}{function to evaluate contribution to
the log posterior from the behavioural model for the k-th track segment.}
\item{\code{fixed}}{a logical vector indicating which locations
should remain fixed.}
\item{\code{x0}}{an array of initial twilight locations.}
\item{\code{time}}{the twilight times.}
\item{\code{rise}}{the sunrise indicators.}
\item{\code{alpha}}{the twilight model parameters.}
\item{\code{beta}}{the behavioural model parameters.}
}
\description{
Curve Model Structures for Essie
}
\details{
Essie requires a model structure that describes the model being
fitted. This function generates basic model structures for curve
fitting metthods that should provide a suitable starting point for
most analyses.

The \code{essieCurveModel} function constructs a model structure
assuming that each twilight profile is associated with a single
location.  The errors in observed log light level are assumed to
be Normally distributed about their expected value.

The properties of the twilight model are determined by
\code{alpha}, which must be a vector of parameters that are
to be applied to each twilight.

The initialization locations \code{x0} are only required when
specifying fixed locations.

Essie assumes that the average speed of travel between successive
locations is Gamma distributed. By default, the speed of travel is
calculated based on the time intervals between the twilights (in
hours), but the intervals of time actually available for travel
can be specified directly with the \code{dt} argument. The
parameters \code{beta[1]} and \code{beta[2]} specify the shape and
rate of the Gamma distribution of speeds.
}
