% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\name{essieMean}
\alias{essieMean}
\alias{essieMode}
\title{Essie locations}
\usage{
essieMean(obj, type = c("full", "forward", "backward"))

essieMode(obj, type = c("full", "forward", "backward"))
}
\arguments{
\item{obj}{a fitted object generated by \code{essie}}

\item{type}{whether to return the full estimate or just the
estimate from the forward or backward pass of the algorithm.}
}
\value{
a list with elements
\item{\code{times}}{the times corresponding to the location estimates.}
\item{\code{x}}{a two column matrix of positions.}
}
\description{
Compute posterior mean or mode locations from an Essie fit
}
\details{
These functions compute the posterior mean or posterior mode
locations for each time point from and Essie fit.
}

