% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lattice.R
\name{essieRaster}
\alias{essieRaster}
\title{Essie probability rasters}
\usage{
essieRaster(obj, k, type = c("full", "forward", "backward", "likelihood"))
}
\arguments{
\item{obj}{a fitted object generated by \code{essie}}

\item{k}{the index of the location}

\item{type}{whether to return the full estimate, the estimate from
the forward or backward pass of the algorithm or the likelihood.}
}
\value{
a raster of probabilities/likelihoods
}
\description{
Return the probabilities from an Essie model as a raster object.
}
\details{
This function returns the likelihood or the posterior probability
of a single location as a raster object.
}

