% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\name{essieThresholdModel}
\alias{essieThresholdModel}
\title{Threshold Model Structures (Essie)}
\usage{
essieThresholdModel(twilight, rise, twilight.model = c("LogNormal", "Gamma",
  "Normal"), alpha, beta, logp0 = function(k, x) 0, x0, fixed = FALSE,
  missing = 0, dt = NULL, zenith = 96)
}
\arguments{
\item{twilight}{the observed times of twilight as POSIXct.}

\item{rise}{logical vector indicating which twilights are sunrise.}

\item{twilight.model}{the model for the errors in twilight times.}

\item{alpha}{parameters of the twilight model.}

\item{beta}{parameters of the behavioural model.}

\item{logp0}{function to evaluate any additional contribution to
the log posterior from the twilight locations.}

\item{x0}{suggested starting points for twilight locations.}

\item{fixed}{logical vector indicating which twilight locations
to hold fixed.}

\item{missing}{integer vector indicating which twilights are missing.}

\item{dt}{time intervals for speed calculation in hours.}

\item{zenith}{the solar zenith angle that defines twilight.}
}
\value{
a list with components
\item{\code{logpk}}{function to evaluate the contributions to the
log posterior from the k-th twilight}
\item{\code{logpbk}}{function to evaluate contribution to
the log posterior from the behavioural model for the k-th track segment.}
\item{\code{fixed}}{a logical vector indicating which locations
should remain fixed.}
\item{\code{x0}}{an array of initial twilight locations.}
\item{\code{time}}{the twilight times.}
\item{\code{rise}}{the sunrise indicators.}
\item{\code{alpha}}{the twilight model parameters.}
\item{\code{beta}}{the behavioural model parameters.}
}
\description{
Threshold Model Structures for Essie
}
\details{
Essie requires a model structure that describes the model being
fitted. This function generates basic model structures for
threshold twilight data that should provide a suitable starting
point for most analyses.

The \code{essieThresholdModel} function constructs a model structure
assuming that each twilight time is associated with a single
location.

One of several models models may be selected for the errors in
twilight times.  The errors in twilight time are defined as the
difference in the observed and true times of twilight, with sign
selected so that a positive error always corresponds to a sunrise
observed after the true time of sunrise, and sunset observed
before the true time of sunset. That is, a positive error
corresponds to the observed light level being lower than expected.

The properties of the twilight model are determined by
\code{alpha}, which must be a vector of parameters that are
to be applied to each twilight.

The \code{twilight.model} argument selects the distribution of the
twilight errors
\describe{
\item{'Normal'}{Normally distributed with mean \code{alpha[,1]} and
standard deviation \code{alpha[,2]},}
\item{'LogNormal'}{Log Normally distributed so the log errors have
mean \code{alpha[,1]} and standard deviation \code{alpha[,2]}, or}
\item{'Gamma'}{Gamma distributed with shape \code{alpha[,1]} and
rate \code{alpha[,2]}.}
}

The initialization locations \code{x0} are only required when
specifying fixed locations.

Essie assumes that the average speed of travel between successive
locations is Gamma distributed. By default, the speed of travel is
calculated based on the time intervals between the twilights (in
hours), but the intervals of time actually available for travel
can be specified directly with the \code{dt} argument. The
parameters \code{beta[1]} and \code{beta[2]} specify the shape and
rate of the Gamma distribution of speeds.

At this point Essie can only deal with purely uninformative
missing values.
}

