% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\name{essieTwilightFreeModel}
\alias{essieTwilightFreeModel}
\title{Twilight Free Model}
\usage{
essieTwilightFreeModel(slices, alpha, beta, logp0 = function(k, x) 0, x0,
  fixed = FALSE, dt = NULL, threshold = 5, zenith = 96)
}
\arguments{
\item{slices}{a list of dataframes of sampled light (see details)}

\item{alpha}{parameters of the obscuration model.}

\item{beta}{parameters of the behavioural model.}

\item{logp0}{function to evaluate any additional contribution to
the log posterior from the daily locations.}

\item{x0}{suggested starting points for twilight locations.}

\item{fixed}{logical vector indicating which twilight locations to
hold fixed.}

\item{dt}{time intervals for speed calculation in hours.}

\item{threshold}{the threshold light level that defines twilight.}

\item{zenith}{the solar zenith angle that defines twilight.}
}
\value{
a list with components
\item{\code{logpk}}{function to evaluate the contributions to the
log posterior from the k-th twilight}
\item{\code{logpbk}}{function to evaluate contribution to
the log posterior from the behavioural model for the k-th track segment.}
\item{\code{fixed}}{a logical vector indicating which locations
should remain fixed.}
\item{\code{x0}}{an array of initial twilight locations.}
\item{\code{time}}{the twilight times.}
\item{\code{alpha}}{the obscuration model parameters.}
\item{\code{beta}}{the behavioural model parameters.}
}
\description{
Twilight Free Model Structure for Essie.
}
\details{
The light data must be given as a list of dataframes where each
dataframe corresponds to approximately one day of data, and
ideally spans the true night.  Each data frame in the list should
contain at least
\tabular{ll}{
\code{Date} \tab observation time (GMT POSIXct) \cr
\code{Light} \tab observed light level \cr
}
It is assumed the dataframes are ordered by time.
}

