% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{estelleMetropolis}
\alias{estelleMetropolis}
\alias{stellaMetropolis}
\title{Metropolis Samplers}
\usage{
estelleMetropolis(model, proposal.x, proposal.z, x0 = NULL, z0 = NULL,
  iters = 1000L, thin = 10L, chains = 1L, verbose = interactive())

stellaMetropolis(model, proposal.x, x0 = NULL, iters = 1000L, thin = 10L,
  chains = 1L, verbose = interactive())
}
\arguments{
\item{model}{a model structure as generated by
\code{thresholdModel}.}

\item{proposal.x}{function for drawing proposals for x.}

\item{proposal.z}{function for drawing proposals for z.}

\item{x0}{Starting values for twilight locations x.}

\item{z0}{Starting values for intermediate locations z.}

\item{iters}{number of samples to draw.}

\item{thin}{rate at which to thin samples.}

\item{chains}{number of chains to sample.}

\item{verbose}{report progress at prompt?}
}
\value{
If there are r samples drawn for each of q chains of p
parameters at n locations, Stella will return a list containing
\item{\code{model}}{the model structure}
\item{\code{x}}{a list of n x p x r arrays of twilight locations
from the q chains}
While in addition Estelle will return
\item{\code{z}}{a list of (n-1) x p x r arrays of intermediate
locations from the q chains}.
}
\description{
Metropolis samplers for Stella or Estelle.
}
\details{
These functions draw samples form posterior for the Stella or
Estelle model by the Metropolis algorithm.
}
\seealso{
\code{\link{thresholdModel}}
}

