% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pimage.R
\name{[.Pimage}
\alias{[.Pimage}
\title{Extract parts of Pimage.}
\usage{
\method{[}{Pimage}(x, i, j, drop = TRUE, ...)
}
\arguments{
\item{x}{Pimage object}

\item{i}{numeric or logical vector}

\item{j}{ignored}

\item{drop}{ignored}

\item{...}{ignored}
}
\value{
RasterLayer, or Pimage, see Details
}
\description{
Extraction with \code{"["} can be done in the usual ways by
numeric or integer indexes, the result is returned as a
\code{\link[raster]{raster}}. There is no replacement method,
attemping to do so is an error.
}
\details{
Use \code{[[.Pimage} to extract a single time step, and
\code{subset.Pimage} to extract multiple elements without coercion
to Raster.
}
\seealso{
\code{\link{cut.Pimage}} for creating temporal
partitions.
}

