% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\name{gcDist}
\alias{gcDist}
\title{Great Circle Distance}
\usage{
gcDist(x1, x2)
}
\arguments{
\item{x1}{a two column matrix of (lon,lat) locations.}

\item{x2}{a two column matrix of (lon,lat) locations.}
}
\value{
vector of interpoint distances (km)
}
\description{
Calculate great circle distancces between points.
}
\details{
Compute the great circle distances from x1 to x2.
}

