% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{gcInterpolate}
\alias{gcInterpolate}
\title{Interpolate Along Great Circles}
\usage{
gcInterpolate(tm, lon, lat, newtm)
}
\arguments{
\item{tm}{times (as POSIXct)}

\item{lon}{vector of longitudes (degrees)}

\item{lat}{vector of latitudes (degrees)}

\item{newtm}{times for which to predict (as POSIXct)}
}
\value{
return a list with components
\item{\code{time}}{the time as POSIXct.}
\item{\code{x}}{a two column matrix of (lon,lat) locations.}
}
\description{
Given sequence of times and locations representing a path,
interpolate along great circle segments to predict locations for a
arbitrary set of times along the path.
}
\details{
Assumes that the original times and locations are sorted in
timewise order with no duplicates.
}

