% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\name{gcOuterDist}
\alias{gcOuterDist}
\title{Great Circle Distance}
\usage{
gcOuterDist(x1, x2)
}
\arguments{
\item{x1}{a two column matrix of (lon,lat) locations.}

\item{x2}{a two column matrix of (lon,lat) locations.}
}
\value{
vector of interpoint distances (km)
}
\description{
Calculate great circle distancces between all pairs of points.
}
\details{
Compute the great circle distances from every point in x1 to every
point is x2, and return the result as a matrix.
}

