% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{geolightConvert}
\alias{geolightConvert}
\title{Convert GeoLight Format}
\usage{
geolightConvert(tFirst, tSecond, type)
}
\arguments{
\item{tFirst}{times of first twilight.}

\item{tSecond}{times of second twilight.}

\item{type}{type of twilight.}
}
\value{
A data frame with columns \item{\code{twilight}}{times of
twilight as POSIXct objects.}  \item{\code{rise}}{logical vector
indicating which twilights are sunrise.}
}
\description{
Convert GeoLight data
}
\details{
This function converts from the tFirst, tSecond format used by
GeoLight to the twilight, rise format used by Stella and Estelle.
}

