% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_coords}
\alias{get_coords}
\title{Get coordinates from a place's name from Google}
\usage{
get_coords(lugares.a.buscar, tiempo.espera = 10)
}
\arguments{
\item{lugares.a.buscar}{What you want to search in Google. Single input or a character vector}

\item{tiempo.espera}{ime measure of how much time you think it's sufficient given your internet connection to load a Google search page and not finding the information,}
}
\value{
dataframe containing the name and the coordinates
}
\description{
Get coordinates from a place's name from Google
}
\examples{
\dontrun{
museo <- get_coords("museo nacional de bellas artes, buenos aires, argentina")
museo
}
}
