% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{gperm}
\alias{gperm}
\title{Generalized Array Permutation}
\usage{
gperm(x, perm)
}
\arguments{
\item{x}{an array.}

\item{perm}{the subscript permutation list.}
}
\value{
The redimensioned array.
}
\description{
Permute and collapse the dimensions of an array
}
\details{
Like \code{aperm} this function permutes the dimensions of an
array, but can also collapse adjacent dimensions.  The permutation
of dimensions is specified by a list of vectors.  The index of
each dimension must appear exactly once, but dimensions that occur
together in a vector are reduced to a single dimension.
}
\examples{
x <- array(1:120,2:5)
dim(gperm(x,list(4,3,2,1)))
dim(gperm(x,list(c(4,3),2,1)))
dim(gperm(x,list(1,c(3,2),4)))
}

