% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_sgat.R
\name{initialization_sgat}
\alias{initialization_sgat}
\title{Automated server creation}
\usage{
initialization_sgat()
}
\value{
"Driver connection"
}
\description{
Automated server creation
}
\examples{
\dontrun{
initialization_sgat()
}
}
