% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raster.R
\name{locationRasterize}
\alias{locationRasterize}
\alias{locationKernelize}
\title{Location Density Raster}
\usage{
locationRasterize(s, grid, weights = 1, zero.is.na = TRUE)

locationKernelize(s, grid, weights = 1, bw = NULL)
}
\arguments{
\item{s}{a single chain or a list of parallel chains generated by
\code{estelleMetropolis} or \code{stellaMetropolis}.}

\item{grid}{raster object that defines the sampling grid}

\item{weights}{weights for each location}

\item{zero.is.na}{if \code{TRUE}, zero counts are returned as \code{NA}.}

\item{bw}{optional bandwidth for the kernel density estimator.}
}
\value{
a raster representing a 2D histogram of locations
}
\description{
Bin locations to form a 2D density raster
}
\details{
Constructs either a 2D histogram or kernel density estimate from
the samples for a sequence of locations.  Optionally, a vector of
weights can be provided to differentially weight samples by
location.
}
