% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raster.R
\name{lonlatFromCell}
\alias{cellFromLonLat}
\alias{lonlatFromCell}
\title{Raster cell longitude and latitudes}
\usage{
lonlatFromCell(raster, cells, spatial = FALSE)

cellFromLonLat(raster, pts)
}
\arguments{
\item{raster}{a raster object}

\item{cells}{the cell numbers}

\item{spatial}{return locations as SpatialPoints object instead of a matrix.}

\item{pts}{the lon, lat locations as a 2 column matrix}
}
\value{
\code{lonlatFromCell} returns the lon,lat locations for
the requested cells as a 2 column matrix, and
\code{cellFromLonLat} returns the cells corresponding to the 2
column matrix of lon,lat positions.
}
\description{
Extract longitude and latitude of raster cells.
}
\details{
Extract the longitude and latitude of the center of the requested
cells of a Raster* object, similar to \code{xyFromCell} and \code{cellFromXY}.
}

