% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{makeTwilightModel}
\alias{makeTwilightModel}
\title{Twilight Error Models}
\usage{
makeTwilightModel(twilight.model = c("Gamma", "LogNormal", "Normal",
  "ModifiedGamma", "ModifiedLogNormal"), alpha)
}
\arguments{
\item{twilight.model}{the model for the errors in twilight times.}

\item{alpha}{parameters of the twilight model.}
}
\value{
a function to evaluate the log density of the twilight
residuals in a threshold model.
}
\description{
Log density of twilight errors
}
\details{
Construct a function to evalute the log density of the twilight
errors in a threshold model.

One of several models models may be selected for the errors in
twilight times.  The errors in twilight time are defined as the
difference in the observed and true times of twilight, with sign
selected so that a positive error always corresponds to a sunrise
observed after the true time of sunrise, and sunset observed
before the true time of sunset. That is, a positive error
corresponds to the observed light level being lower than expected.

The properties of the twilight model are determined by
\code{alpha}, which must be either a vector of parameters that are
to be applied to each twilight, or a matrix of parameters with one
row for each twilight.

The \code{twilight.model} argument selects the distribution of the
twilight errors
\describe{
\item{'Normal'}{Normally distributed with mean \code{alpha[,1]} and
standard deviation \code{alpha[,2]},}
\item{'LogNormal'}{Log Normally distributed so the log errors have
mean \code{alpha[,1]} and standard deviation \code{alpha[,2]}, or}
\item{'Gamma'}{Gamma distributed with shape \code{alpha[,1]} and
rate \code{alpha[2]}.}
}
The 'LogNormal' and 'Gamma' models forbid negative errors, that
is, the observed light cannot be brighter than expected.  There
are modified variants of these models for which negative errors
are extremely unlikely, but not forbidden, and can be used to
generate suitable initialization locations for their unmodified
counterparts.
}

