% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{mvnorm}
\alias{bmvnorm}
\alias{mvnorm}
\title{Multivariate Normal Samples}
\usage{
mvnorm(S, s = 1, n = 1, tol = 1e-06)

bmvnorm(S, m, s = 1)
}
\arguments{
\item{S}{a covariance matrix or an array of covariance matrices.}

\item{s}{a scale factor applied to S.}

\item{n}{number of deviates to draw.}

\item{tol}{minimum allowable variance.}

\item{m}{dimension of each deviate.}
}
\value{
A function that draws bivariate Normal deviates with mean
given by its first argument.
}
\description{
Construct a sampler to draw multivariate Normal deviates.
}
\details{
Construct a sampler that draws Multivariate Normal deviates with
covariances determined by \code{S} and mean determined by its
first argument.

The \code{mvnorm} function constructs a function that generates
\code{n} independent Multivariate Normal deviates, where the
covariance of each deviate is specified by \code{S} which must be
either a \code{m}x\code{m} covariance matrix or an
\code{n}x\code{m}x\code{m} array of covariance matrices.

The \code{bmvnorm} constructs a function that generates \code{n}
correlated Multivariate Normal deviates, where the joint
covariance is specified by \code{S} which must be a
\code{nm}x\code{nm} covariance matrix (as generated by
\code{chainBcov}).
}

