% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{nlocation}
\alias{nlocation}
\title{Number of locations}
\usage{
nlocation(s)
}
\arguments{
\item{s}{an array or a list of arrays.}
}
\value{
size of the first dimension of the array.
}
\description{
Number of locations
}
\details{
A convience function to determine the number of locations a chain,
or set of initial locations or a location summary
describe. Assumes \code{s} is either an array or a list of arrays
in which the first dimension corresponds to location, and returns
the length of the first dimension.
}
