% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{parametersGamma}
\alias{parametersGamma}
\title{Alternate Gamma Parametrization}
\usage{
parametersGamma(mean, sd)
}
\arguments{
\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}
}
\value{
Returns a 2 column array of the shape and rate parameters
that will yield the required mean and standard deviation.
}
\description{
Calculate shape and rate parameters of the Gamma distribution
}
\details{
The Gamma distribution is usually parameterized in terms of the
shape and rate parameters.  The \code{parametersGamma} function
deterimines the shape and rate parameters that will yield a Gamma
distribution with a desired mean and standard deviation.
}
\examples{
## Shape and rate that give a mean of 3 and sd of 2
parametersGamma(3,2)
}

