% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pimage.R
\name{print.Pimage}
\alias{as.list.Pimage}
\alias{c.Pimage}
\alias{length.Pimage}
\alias{plot.Pimage}
\alias{print}
\alias{print.Pimage}
\alias{str.Pimage}
\title{General methods for Pimage}
\usage{
\method{print}{Pimage}(x, ...)

\method{str}{Pimage}(object, ...)

\method{length}{Pimage}(x, ...)

\method{c}{Pimage}(..., recursive = FALSE)

\method{as.list}{Pimage}(x, ...)

\method{plot}{Pimage}(x, ...)
}
\arguments{
\item{x}{Pimage}

\item{...}{arguments passed along}

\item{object}{Pimage}

\item{recursive}{ignored}
}
\description{
These are the basics for print, plot, concatenation, coercion.
}

