% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pimage.R
\name{[<-.Pimage}
\alias{[<-.Pimage}
\title{Sub element replacement is disallowed}
\usage{
\method{[}{Pimage}(x, \dots) <- value
}
\arguments{
\item{x}{Pimage}

\item{...}{ignored}

\item{value}{ignored}
}
\description{
Sub element replacement is disallowed
}

