% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{satelliteModel}
\alias{satelliteModel}
\alias{satelliteModel0}
\title{Satellite Model Structures}
\usage{
satelliteModel(
  time,
  X,
  location.model = c("Normal", "T"),
  sd,
  df = NULL,
  beta,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  x0,
  z0 = NULL,
  fixedx = FALSE,
  dt = NULL
)

satelliteModel0(
  time,
  X,
  location.model = c("Normal", "T"),
  sd,
  df = NULL,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  fixedx = FALSE
)
}
\arguments{
\item{time}{the times of the satellite determined locations.}

\item{X}{the satellite determined locations.}

\item{location.model}{the model for the errors in satellite
locations.}

\item{sd}{a vector or two column matrix of dispersions for the
location model.}

\item{df}{a vector or two column matrix of degrees of freedom for
the t location model.}

\item{beta}{parameters of the behavioural model.}

\item{logp.x}{function to evaluate any additional contribution to
the log posterior from the satellite estimated locations.}

\item{logp.z}{function to evaluate any additional contribution to
the log posterior from the intermediate locations.}

\item{x0}{suggested starting points for the satellite locations.}

\item{z0}{suggested starting points for intermediate locations.}

\item{fixedx}{logical vector indicating which satellite locations
to hold fixed.}

\item{dt}{time intervals for speed calculation in hours.}
}
\value{
a list with components
\item{\code{logpx}}{function to evaluate the contributions to the
log posterior from the twilight model}
\item{\code{logpz}}{function to evaluate the contributions to the
log posterior from the prior for the z locations}
\item{\code{estelle.logpb}}{function to evaluate contribution to
the log posterior from the behavioural model for estelle.}
\item{\code{stella.logpb}}{function to evaluate contribution to
the log posterior from the behavioural model for stella.}
\item{\code{fixedx}}{a logical vector indicating which locations
should remain fixed.}
\item{\code{x0}}{an array of initial twilight locations.}
\item{\code{z0}}{an array of initial intermediate locations.}
\item{\code{time}}{the times of the satellite determined
locations.}
\item{\code{X}}{the satellite determined locations.}
}
\description{
Satellite Model Structure for Stella and Estelle
}
\details{
Stella and Estelle require a model structure that describes the
model being fitted. This function generates a basic model
structure for satellite data that should provide a suitable
starting point for most analyses.

The \code{satelliteModel} function constructs a model structure
assuming that assumes the locations \code{X} determined by the
satellite are independently distributed about the corresponding
true locations.  The \code{location.model} parameter selects
whether \code{(X-x)/sd} is

\describe{
\item{'Normal'}{Normally distributed with zero mean and
unit variance, or}
\item{'T'}{t distributed with degrees of freedom df.}
}

Both Estelle and Stella variants of the model assume that the
average speed of travel between successive locations is Gamma
distributed, and for Estelle models, the change in bearing
(degrees) along the dog-leg path segments can be assumed Normally
distributed with mean zero.  By default, the speed of travel is
calculated based on the time intervals between the twilights (in
hours), but the intervals of time actually available for travel
can be specified directly with the \code{dt} argument.

If \code{beta} is a vector, then \code{beta[1]} and \code{beta[2]}
specify the shape and rate of the Gamma distribution of speeds.
If \code{beta} has three elements, then \code{beta[3]} specifies
the standard deviation of the change in bearing (in degrees) along
dog leg paths.

The \code{satelliteModel0} function constructs the non-movement
elements of the model, and the movement elements of the model are
constructed by the \code{speedGammaModel} function.
}
