% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\docType{package}
\name{SGAT-package}
\alias{SGAT-package}
\title{Solar/Satellite Geolocation for Animal Tracking}
\description{
Provides facilities for estimating broadscale animal motions from
archival or satellite tag data, similar to the \pkg{GeoLight} and
\pkg{tripEstimation} packages.
}
\author{
S. Wotherspoon, M. Sumner and S. Lisovski.
}

