% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{SGAT2Movebank}
\alias{SGAT2Movebank}
\title{Summaries of Location Samples for Movebank}
\usage{
SGAT2Movebank(s, time = NULL, group = NULL, discard = 0,
  alpha = 0.95, chains = NULL)
}
\arguments{
\item{s}{a single chain or a list of parallel chains generated by
\code{estelleMetropolis} or \code{stellaMetropolis}.}

\item{time}{the times corresponding to the x locations.}

\item{group}{integer vector that defines the twilight groups.  If
code{group[k]==j} then the k-th twilight occurs at location j. Default is \code{NULL} for non-grouped models.}

\item{discard}{number of initial samples to discard.}

\item{alpha}{coverage of the credible intervals calculated by
\code{locationSummary}.}

\item{chains}{the set of chains to retain, or \code{NULL}.}
}
\value{
\item{\code{SGAT2Movebank}}{returns a dataframe of summary quantities for each location or grouped location.}
}
\description{
Summarize a set of location samples in the format required for upload to Movebank
}
\details{
These functions compute various summaries of a sample or list of
samples generated by \code{estelleMetropolis} or
\code{stellaMetropolis}.

These functions accept either a sample from a single mcmc run, or
a list of samples from parallel mcmc runs.
}
