% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\docType{data}
\name{Shearwater}
\alias{Shearwater}
\title{Short-tailed Shearwater Track}
\format{
A data frame with 3 columns and 72 rows.  The columns
represent
\tabular{rl}{
\code{twilight} \tab times of twilight. \cr
\code{rise} \tab logical indicating which twilights are sunrise. \cr
\code{burrow} \tab logical indicating when the bird in its burrow.
}
}
\description{
Twilight times from a flight of a Short-tailed Shearwater.
}
\details{
Twilight times derived from light intensity measurements from an
archival tag on a Short-tailed Shearwater (\emph{Puffinus
tenuirostris}).  The bird was tagged at its burrow on Wedge
Island, Tasmania Australia (147.670E 43.133S).  The bird makes two
foraging trips, returning to its burrow after each trip.

When the bird is in its burrow, no light is recorded by the tag
and the twilights reported in the data set are the calculated
twilights Wedge Island when solar zenith is 95 degrees.

These data supplied courtesy of Delphi Ward and Mark Hindell,
Institute of Marine and Antarctic Studies, University of
Tasmania.
}
\keyword{data}
