% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raster.R
\name{slice}
\alias{slice}
\alias{sliceIndices}
\alias{sliceInterval}
\alias{slices}
\title{Location Density Estimates}
\usage{
slice(slices, k, mcmc = slices$mcmc, grid = slices$grid,
  weights = slices$weights, method = slices$method,
  bandwidth = slices$bandwidth, chains = NULL)

slices(type = c("primary", "intermediate"), breaks = NULL, mcmc = NULL,
  grid = raster(), weights = NULL, method = c("hist", "kde"),
  bandwidth = 0, include.lowest = TRUE, right = FALSE)

sliceInterval(slices, k, mcmc = slices$mcmc)

sliceIndices(slices, mcmc = slices$mcmc)
}
\arguments{
\item{slices}{an object generated by \code{slices}}

\item{k}{the index of the time slice to bin.}

\item{mcmc}{object generated by generated by
\code{estelleMetropolis} or \code{stellaMetropolis}.}

\item{grid}{raster object that defines the sampling grid.}

\item{weights}{weights for each location.}

\item{method}{return a kernel density estimate (\code{"kde"})
or simple histogram (\code{"hist"})?}

\item{bandwidth}{bandwidth of the kernel density estimator.}

\item{chains}{NULL or the subset of chains to bin.}

\item{type}{construct density rasters for \code{primary} (x) or
\code{intermediate} (z) locations}

\item{breaks}{NULL, or a specification suitable for
\code{cut.POSIXt} to define the time slices to bin.}

\item{include.lowest}{parameter to \code{cut.POSIXt}.}

\item{right}{parameter to \code{cut.POSIXt}.}
}
\value{
\code{slice} returns the locations for time slice of the
track binned into a raster, \code{slices} returns a slices
object that defines the time slices into which to bin,
\code{sliceInterval} returns the time interval spanned by a
slice, and \code{sliceIndices} returns the valid set of indices
that will yield a raster.
}
\description{
Generate density estimates for time slices of a trip
}
\details{
These functions divide a trip into time slices, and generate
location density estimates for each time slice.

The \code{slices} function defines the slices into which the trip
is divided, and whether density estimates are generated for the
primary or intermediate locations. If \code{breaks} is NULL, each
location forms a separate time slice, otherwise \code{breaks}
divides the trip into time slices in the same style as
\code{\link{cut.POSIXt}}.  A default set of samples and a raster
defining the spatial grid may also be specified.

The \code{slice} function generates a density estimate for a time
slice, \code{sliceInterval} returns the corresponding time
interval, and \code{sliceIndices} returns the indices that will
yield non null raster.
}

