% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{speedGammaModel}
\alias{speedGammaModel}
\title{Gamma Behavioural Model}
\usage{
speedGammaModel(beta, dt)
}
\arguments{
\item{beta}{parameters of the behavioural model.}

\item{dt}{time intervals for speed calculation in hours.}
}
\value{
Functions to evaluate the contributions to the log
posterior for the Estelle and Stella models.
}
\description{
Movement model that assumes speeds are Gamma distributed
}
\details{
This function implements a movement model that assumes the speed
of travel between locations is Gamma distributed, and for Estelle
models, the change in bearing along the dog-leg path segments can
be assumed Normally distributed with mean zero.

For Stella models, average speeds is calculated along great circle
paths between primary locations (x).  For Estelle, average speed
is calculated along dog leg paths through the intermediate points
(z), and the change in bearing at each intermediate point calculated.

If \code{beta} is a vector, then \code{beta[1]} and \code{beta[2]}
specify the shape and rate of the Gamma distribution of speeds.
If \code{beta} has three elements, then \code{beta[3]} specifies
the standard deviation of the change in bearing (in degrees) along
dog leg paths.

Alternately, these parameters can be specified individually for
each track segment by passing \code{beta} as a matrix with one row
for each segment.
}
\seealso{
\code{\link{satelliteModel}}, \code{\link{thresholdModel}},
\code{\link{groupedThresholdModel}}, \code{\link{curveModel}}.
}

