% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{thresholdEstimate}
\alias{thresholdEstimate}
\alias{thresholdLocation}
\alias{thresholdPath}
\title{Simple Threshold Geolocation Estimates}
\usage{
thresholdEstimate(trise, tset, zenith = 96, tol = 0)

thresholdLocation(twilight, rise, zenith = 96, tol = 0.08)

thresholdPath(twilight, rise, time = twilight, zenith = 96, tol = 0.08,
  unfold = TRUE)
}
\arguments{
\item{trise}{vector of sunrise times.}

\item{tset}{vector of sunset times.}

\item{zenith}{the solar zenith angle that defines twilight.}

\item{tol}{tolerance on the sine of the solar declination.}

\item{twilight}{the observed times of twilight as POSIXct.}

\item{rise}{logical vector indicating which twilights are sunrise.}

\item{time}{times for which locations are required.}

\item{unfold}{if \code{TRUE}, unfold longitudes across the dateline.}
}
\value{
\code{thresholdEstimate} returns estimated locations as a
two column (lon,lat) matrix.  \code{thresholdLocation} and
\code{thresholdPath} return a list with components
\item{\code{time}}{the time as POSIXct.}
\item{\code{x}}{a two column matrix of (lon,lat) locations.}
}
\description{
Estimate location from consecutive twilights
}
\details{
These functions estimate the location of a stationary observer
given the times at which the observer sees two successive
twilights. \code{thresholdEstimate} estimates locations given
pairs of times of sunrise and sunset. \code{thresholdLocation}
is a wrapper for \code{thresholdEstimate} that estimates
locations given a sequence twilight times and rise indicators,
while \code{thresholdPath} interpolates the estimates generated
by \code{thresholdLocation} to give locations at a sequence of
arbitrary times.

Longitude is estimated by computing apparent time of local noon
from sunrise and sunset, and determining the longitude for which
this is noon.  Latitude is estimated from the required zenith and
the sun's hour angle for both sunrise and sunset, and averaged.

When the solar declination is near zero (at the equinoxes)
latitude estimates are extremely sensitive to errors.  Where the
sine of the solar declination is less than \code{tol}, the
latitude estimates are returned as \code{NA}.

The \code{thresholdPath} function interpolates the estimates
generated by \code{thresholdLocation} to produce estimates at the
arbitrary set of times specified the by the \code{time} argument.
If \code{time} is \code{NULL}, \code{thresholdPath} returns the
estimates generated by \code{thresholdLocation}. If
\code{unfold=TRUE}, \code{thresholdPath} attempts to construct a
continuous path that does not wrap longitudes into
(-180,180]. However, this process can fail if the observer crosses
the dateline near an equinox, and it may be necessary to make
manual adjustments.

These functions provides the same basic functionality of the
\code{coord} function from \pkg{GeoLight}, but are based on
different astronomical approximations.
}
\seealso{
\code{\link{zenith}}
}

