% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{thresholdLocation}
\alias{thresholdLocation}
\alias{thresholdLocationRS}
\alias{thresholdPath}
\title{Simple Threshold Geolocation Estimates}
\usage{
thresholdLocation(twilight, rise, zenith = 96, tol = 0.02)

thresholdLocationRS(trise, tset, zenith = 96, tol = 0)

thresholdPath(twilight, rise, time = twilight, zenith = 96, tol = 0.02,
  unfold = TRUE)
}
\arguments{
\item{twilight}{the observed times of twilight as POSIXct.}

\item{rise}{logical vector indicating which twilights are sunrise.}

\item{zenith}{the solar zenith angle that defines twilight.}

\item{tol}{tolerance on the sine of the solar declination.}

\item{trise}{vector of sunrise times.}

\item{tset}{vector of sunset times.}

\item{time}{times for which locations are required.}

\item{unfold}{if \code{TRUE}, unfold longitudes across the dateline.}
}
\value{
Return a list with components
\item{\code{time}}{the time as POSIXct.}
\item{\code{x}}{a two column matrix of (lon,lat) locations.}
}
\description{
Estimate locations from consecutive twilights
}
\details{
These functions estimate the location of a stationary observer
given the times at which the observer sees two successive
twilights. \code{thresholdLocationRS} estimates locations given
pairs of times of sunrise and sunset. \code{thresholdLocation}
is a wrapper for \code{thresholdLocationRS} that estimates
locations given a sequence twilight times and rise indicators,
while \code{thresholdPath} interpolates the estimates generated
by \code{thresholdLocation} to give locations at a sequence of
arbitrary times.

Longitude is estimated by computing apparent time of local noon
from sunrise and sunset, and determining the longitude for which
this is noon.  Latitude is estimated from the required zenith and
the sun's hour angle for both sunrise and sunset, and averaged.

When the solar declination is near zero (at the equinoxes)
latitude estimates are extremely sensitive to errors.  Where the
sine of the solar declination is less than \code{tol}, the
latitude estimates are returned as \code{NA}.

The \code{thresholdPath} function interpolates the estimates
generated by \code{thresholdLocation} to produce estimates at the
arbitrary set of times specified the by the \code{time} argument.
If \code{time} is \code{NULL}, \code{thresholdPath} returns the
estimates generated by \code{thresholdLocation}. If
\code{unfold=TRUE}, \code{thresholdPath} attempts to construct a
continuous path that does not wrap longitudes into
(-180,180]. However, this process can fail if the observer crosses
the dateline near an equinox, and it may be necessary to make
manual adjustments.

These functions provides the same basic functionality of the
\code{coord} function from \pkg{GeoLight}, but are based on
different astronomical approximations.
}
\examples{
## Simulate twilight times for January 150E, -40S
tm <- rep(seq(as.POSIXct("2016-01-01","GMT"),by="day",length.out=3),each=2)
rise <- rep(c(FALSE,TRUE),3)
twl <- twilight(tm,150,-40,rise)
## thresholdLocation - gives locations for approx local noon and midnight
(loc <- thresholdLocation(twl,rise))
## Sunsets and sunrises for estimated locations
data.frame(simulated=twl[c(1,3,3,5,5)],
           estimated=twilight(loc$time,loc$x[,1],loc$x[,2],FALSE))
data.frame(simulated=twl[c(2,2,4,4,6)],
           estimated=twilight(loc$time,loc$x[,1],loc$x[,2],TRUE))
## thresholdPath - interpolates to give locations at twilight times
(loc <- thresholdPath(twl,rise))
data.frame(simulated=twl,
           estimated=twilight(loc$time,loc$x[,1],loc$x[,2],rise))

## Near the equinox latitude estimation degrades and can fail
tm <- rep(seq(as.POSIXct("2016-03-16","GMT"),by="day",length.out=3),each=2)
rise <- rep(c(FALSE,TRUE),3)
twl <- twilight(tm,150,-40,rise)
thresholdLocation(twl,rise)

## Estimates degrade if the tag is moving
## Simulate twilight times for January 150E, -40S
tm <- rep(seq(as.POSIXct("2016-01-01","GMT"),by="day",length.out=3),each=2)
rise <- rep(c(FALSE,TRUE),3)
twl <- twilight(tm,seq(150,by=0.8,length=6),-40,rise)
(loc <- thresholdLocation(twl,rise))
## Sunsets and sunrises for estimated locations
data.frame(simulated=twl[c(1,3,3,5,5)],
           estimated=twilight(loc$time,loc$x[,1],loc$x[,2],FALSE))
data.frame(simulated=twl[c(2,2,4,4,6)],
           estimated=twilight(loc$time,loc$x[,1],loc$x[,2],TRUE))
}
\seealso{
\code{\link{zenith}}
}

