% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{thresholdModel}
\alias{thresholdModel}
\alias{thresholdModel0}
\alias{groupedThresholdModel}
\alias{groupedThresholdModel0}
\title{Threshold Model Structures}
\usage{
thresholdModel(
  twilight,
  rise,
  twilight.model = c("Gamma", "LogNormal", "Normal", "ModifiedGamma",
    "ModifiedLogNormal"),
  alpha,
  beta,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  x0,
  z0 = NULL,
  fixedx = FALSE,
  missing = 0,
  dt = NULL,
  zenith = 96
)

thresholdModel0(
  twilight,
  rise,
  twilight.model = c("Gamma", "LogNormal", "Normal", "ModifiedGamma",
    "ModifiedLogNormal"),
  alpha,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  fixedx = FALSE,
  missing = 0,
  zenith = 96
)

groupedThresholdModel(
  twilight,
  rise,
  group,
  twilight.model = c("Gamma", "LogNormal", "Normal", "ModifiedGamma",
    "ModifiedLogNormal"),
  alpha,
  beta,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  x0,
  z0 = NULL,
  fixedx = FALSE,
  missing = 0,
  dt = NULL,
  zenith = 96
)

groupedThresholdModel0(
  twilight,
  rise,
  group,
  twilight.model = c("Gamma", "LogNormal", "Normal", "ModifiedGamma",
    "ModifiedLogNormal"),
  alpha,
  logp.x = function(x) rep.int(0L, nrow(x)),
  logp.z = function(z) rep.int(0L, nrow(z)),
  fixedx = FALSE,
  missing = 0,
  zenith = 96
)
}
\arguments{
\item{twilight}{the observed times of twilight as POSIXct.}

\item{rise}{logical vector indicating which twilights are sunrise.}

\item{twilight.model}{the model for the errors in twilight times.}

\item{alpha}{parameters of the twilight model.}

\item{beta}{parameters of the behavioural model.}

\item{logp.x}{function to evaluate any additional contribution to
the log posterior from the twilight locations.}

\item{logp.z}{function to evaluate any additional contribution to
the log posterior from the intermediate locations.}

\item{x0}{suggested starting points for twilight locations.}

\item{z0}{suggested starting points for intermediate locations.}

\item{fixedx}{logical vector indicating which twilight locations
to hold fixed.}

\item{missing}{integer vector indicating which twilights were
unobserved and why.}

\item{dt}{time intervals for speed calculation in hours.}

\item{zenith}{the solar zenith angle that defines twilight.}

\item{group}{integer vector that defines the twilight groups.  If
code{group[k]==j} then the k-th twilight occurs at location j.}
}
\value{
a list with components
\item{\code{logpx}}{function to evaluate the contributions to the
log posterior from the twilight model}
\item{\code{logpz}}{function to evaluate the contributions to the
log posterior from the prior for the z locations}
\item{\code{estelle.logpb}}{function to evaluate contribution to
the log posterior from the behavioural model for estelle.}
\item{\code{stella.logpb}}{function to evaluate contribution to
the log posterior from the behavioural model for stella.}
\item{\code{residuals}}{function to evaluate the twilight model
residuals.}
\item{\code{fixedx}}{a logical vector indicating which locations
should remain fixed.}
\item{\code{x0}}{an array of initial twilight locations.}
\item{\code{z0}}{an array of initial intermediate locations.}
\item{\code{time}}{the twilight times.}
\item{\code{rise}}{the sunrise indicators.}
\item{\code{group}}{the grouping vector.}
}
\description{
Threshold Model Structures for Stella and Estelle
}
\details{
Stella and Estelle require a model structure that describes the
model being fitted. These function generate basic model structures
for threshold twilight data that should provide a suitable
starting point for most analyses.

The \code{thresholdModel} function constructs a model structure
assuming that each twilight time is associated with a single
location, while the \code{groupedThresholdModel} function allows
multiple twilight times to be associated with a single location.

One of several models models may be selected for the errors in
twilight times.  The errors in twilight time are defined as the
difference in the observed and true times of twilight, with sign
selected so that a positive error always corresponds to a sunrise
observed after the true time of sunrise, and sunset observed
before the true time of sunset. That is, a positive error
corresponds to the observed light level being lower than expected.

The properties of the twilight model are determined by
\code{alpha}, which must be either a vector of parameters that are
to be applied to each twilight, or a matrix of parameters with one
row for each twilight.

The \code{twilight.model} argument selects the distribution of the
twilight errors
\describe{
\item{'Normal'}{Normally distributed with mean \code{alpha[,1]} and
standard deviation \code{alpha[,2]},}
\item{'LogNormal'}{Log Normally distributed so the log errors have
mean \code{alpha[,1]} and standard deviation \code{alpha[,2]}, or}
\item{'Gamma'}{Gamma distributed with shape \code{alpha[,1]} and
rate \code{alpha[2]}.}
}
The 'LogNormal' and 'Gamma' models forbid negative errors, that
is, the observed light cannot be brighter than expected.  There
are modified variants of these models for which negative errors
are extremely unlikely, but not forbidden, and can be used to
generate suitable initialization locations for their unmodified
counterparts.

The initialization locations \code{x0} and \code{z0} must be
consistent with the chosen twilight model.  That is, if
'LogNormal' or 'Gamma' models are selected, the \code{x0} cannot
yield negative twilight errors.

Both Estelle and Stella variants of the model assume that the
average speed of travel between successive locations is Gamma
distributed, and for Estelle models, the change in bearing
(degrees) along the dog-leg path segments can be assumed Normally
distributed with mean zero.  By default, the speed of travel is
calculated based on the time intervals between the twilights (in
hours), but the intervals of time actually available for travel
can be specified directly with the \code{dt} argument.

If \code{beta} is a vector, then \code{beta[1]} and \code{beta[2]}
specify the shape and rate of the Gamma distribution of speeds.
If \code{beta} has three elements, then \code{beta[3]} specifies
the standard deviation of the change in bearing (in degrees) along
dog leg paths.

Twilights can be missing because either the light record was too
noisy at that time to estimate twilight reliably, or because the
tag was at very high latitude and no twilight was observed.
Missing twilights should be replaced with an approximate time of
twilight, and the vector \code{missing} used to indicate which
twilights are approaximate and which are true.  This should be a
vector of integers, one for each twilight where the integer codes
signify
\describe{
\item{0:}{The twilight is not missing.}
\item{1:}{The twilight is missing, but a twilight did occur.}
\item{2:}{The twilight is missing because twilight did not occur.}
\item{3:}{The twilight is missing and it is not known if a twilight occurred.}
}

The \code{thresholdModel0} and \code{groupedThresholdModel0}
functions construct the non-movement elements of the model, and
the movement elements of the model are constructed by the
\code{speedGammaModel} function.
}
