% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{thresholdSensitivity}
\alias{thresholdSensitivity}
\title{Threshold Geolocation Sensitivity}
\usage{
thresholdSensitivity(
  rise,
  set,
  zenith = 96,
  range = 100,
  sr.mulog,
  sr.sdlog,
  ss.mulog,
  ss.sdlog,
  sr.proposal,
  ss.proposal,
  n.thin = 10,
  n.iters = 1000
)
}
\arguments{
\item{rise}{observed time of sunrise as POSIXct.}

\item{set}{observed time of sunset as POSIXct.}

\item{zenith}{the solar zenith angle that defines twilight.}

\item{range}{maximum range of travel between twilights (km).}

\item{sr.mulog}{log mean parameter for the Log Normal distribution
of sunrise errors.}

\item{sr.sdlog}{log standard deviation parameter for the Log
Normal distribution of sunrise errors.}

\item{ss.mulog}{log mean parameter for the Log Normal distribution
of sunset errors.}

\item{ss.sdlog}{log standard deviation parameter for the Log
Normal distribution of sunset errors.}

\item{sr.proposal}{function for drawing from the proposal
distribution for sunrise location.}

\item{ss.proposal}{function for drawing from the proposal
distribution for sunrise location.}

\item{n.thin}{rate at which to thin samples.}

\item{n.iters}{total number of samples to draw.}
}
\value{
a list with three components
\item{\code{p0}}{the threshold estimate}
\item{\code{rise}}{the sampled sunrise locations as a two column
matrix}
\item{\code{set}}{the sampled sunset locations as a two column
matrix}
}
\description{
Estimate locations by the threshold method assuming a
nonstationary observer and errors in estimated twilights
}
\details{
Given the times of a single sunrise and sunset pair,
\code{thresholdSensitivity} estimates the location of the tagged
animal at sunrise and at sunset assuming that during this time the
animal moves no further than a given maximum range, and that the
observed times of sunrise and sunset contain an additive log
Normally distributed error with known mean and variance. These
errors are directed so that observed sunrise occurs earlier than
true sunrise, and the observed sunset occurs later than true
sunrise.

\code{thresholdSensitivity} implements a Metropolis sampler to
draw samples from the posterior distribution for the sunrise and
sunset.
}
