% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{trackDist}
\alias{trackDist}
\alias{trackDist2}
\title{Distance along a path}
\usage{
trackDist(x)

trackDist2(x, z)
}
\arguments{
\item{x}{a two column matrix of (lon,lat) locations along the path.}

\item{z}{a two column matrix of (lon,lat) intermediate locations
along the path.}
}
\value{
vector of interpoint distances (km)
}
\description{
Distances along a path
}
\details{
The \code{trackDist} computes the great circle distances (in km)
between successive locations along path. The \code{trackDist2}
accepts a second sequence of intermediate points, and computes the
great circle distances along the dog leg paths from \code{x[i,]}
to \code{z[i,]} to \code{x[i+1,]}.
}

