% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{trackMidpts}
\alias{trackMidpts}
\title{Path Midpoints}
\usage{
trackMidpts(p, fold = FALSE)
}
\arguments{
\item{p}{a two column matrix of (lon,lat) locations along the
path.}

\item{fold}{should the longitudes be folded into [-180,180).}
}
\value{
a two column matrix of (lon,lat) midpoints.
}
\description{
Compute the midpoints of the great circle segments between the
locations along a path.
}

