% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripadvisor_places.R
\name{tripadvisor_places}
\alias{tripadvisor_places}
\title{Search for the n most popular places in a city according to Tripadvisor}
\usage{
tripadvisor_places(ciudad, n.resultados = Inf)
}
\arguments{
\item{ciudad}{City or area where you want the places' information}

\item{n.resultados}{Maximum number of results to retrieve. If not specified, all results will be retrieved}
}
\value{
Character vector with the names of the most popular places of the searched city or area according to Tripadvisor
}
\description{
Search for the n most popular places in a city according to Tripadvisor
}
\examples{

\dontrun{
tripadvisor_places("Pinamar, Argentina", 10)
}
}
