% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SGAT.R
\name{twilightPairs}
\alias{twilightPairs}
\title{Extract sunrise/sunset pairs}
\usage{
twilightPairs(twilight, rise)
}
\arguments{
\item{twilight}{the observed times of twilight as POSIXct.}

\item{rise}{logical vector indicating which twilights are sunrise.}
}
\value{
A dataframe with columns
\item{\code{Twilight1}}{times of earlier twilight as POSIXct objects}
\item{\code{Twilight2}}{times of later twilight as POSIXct objects}
\item{\code{Day}}{logical vector indicating whether the twilights span a day.}
\item{\code{Mid}}{the midpont of the two twilights.}
}
\description{
Convert streams of twilights to sunrise/sunset pairs
}
\details{
This function converts the twilight, rise format used by Stella
and Estelle into successive sunrise and sunset pairs.
}

