% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raster.R
\name{twilightResidualsMap}
\alias{twilightResidualsMap}
\title{Twilight Residuals}
\usage{
twilightResidualsMap(twilight, rise, grid, zenith = 96)
}
\arguments{
\item{twilight}{an observed time of twilight}

\item{rise}{\code{TRUE} if twilight is a sunrise}

\item{grid}{raster object that define the sampling grid.}

\item{zenith}{the solar zenith angle that defines twilight.}
}
\value{
a raster of twilight residuals (in minutes).
}
\description{
Spatial maps of twilight residuals
}
\details{
This function calculates the twilight residuals corresponding to
a single observed twilight across a grid of locations.
}
\seealso{
\code{\link{twilightResiduals}}
}

