% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{twilightSolartime}
\alias{twilightSolartime}
\title{Solar Time of Sunrise and Sunset}
\usage{
twilightSolartime(solar, lon, lat, rise, zenith = 96)
}
\arguments{
\item{solar}{output of \code{solar} for approximate times of
twilight.}

\item{lon}{vector of longitudes.}

\item{lat}{vector of latitudes.}

\item{rise}{logical vector indicating whether to compute rise or
set.}

\item{zenith}{the solar zenith angle that defines twilight.}
}
\value{
a vector of twilight times in solar time (degrees)
}
\description{
Estimate time of sunrise or sunset for a given location given the
approximate solar time of twilight
}
\details{
Solar declination and equation of time vary slowly over the day,
and so the values of the Solar declination and equation of time at
sunrise/sunset ca be caclulated approximately is an approximate
time of sunrise/sunset is known. The sun's hour angle and hence
sunrise/sunset for the required zenith can then be calculated from
these approximations.

Note this function returns the time of twilight in solar time.
}
\seealso{
\code{\link{twilight}}
}

