% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mobility_data.R
\name{visualize_mobility_data}
\alias{visualize_mobility_data}
\title{Shiny visualization tool for dowloaded Google mobility data}
\usage{
visualize_mobility_data(mobility_data)
}
\arguments{
\item{mobility_data}{Data frame containing raw Google mobility data downloaded with \code{\link{mobility_var}} function}
}
\value{
Opens a Shiny app plotting the Google mobilty data in different ways
}
\description{
Shiny visualization tool for dowloaded Google mobility data
}
\examples{
mobi <- mobility_var(c("AR", "BR", "CL"), c(2020, 2021))
visualize_mobility_data(mobi)
}
