% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{wrapLon}
\alias{wrapLon}
\alias{unwrapLon}
\title{Wrap Locations Around the Dateline.}
\usage{
wrapLon(lon, lmin = -180)

unwrapLon(lon, lmin = -180)
}
\arguments{
\item{lon}{a vector of longitudes}

\item{lmin}{western boundary for wrapped longitudes}
}
\value{
a vector of longitudes
}
\description{
These functions wrap and unwrap a sequence of longitudes around
the dateline.
}
\details{
The \code{wrapLon} function wraps the longitudes back into the
interval [lmin,lmin+360).  The \code{unwrapLon} function unwraps a
sequence of longitudes so the the initial point lies in
[lmin,lmin+360), but the subsequent longitudes in the sequence may
wander outside that range.
}
