% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGAT.R
\name{zenithSimulate}
\alias{twilightPerturb}
\alias{twilightSimulate}
\alias{zenithSimulate}
\title{Solar Zenith and Twilight Simulation}
\usage{
zenithSimulate(tm, lon, lat, tm.out)

twilightSimulate(dfz, zenith = 96)

twilightPerturb(dft, err)
}
\arguments{
\item{tm}{vector of times that specify the template track.}

\item{lon}{vector of longitude that specify the template track.}

\item{lat}{vector of latitude that specify the template track.}

\item{tm.out}{vector of times to which the template is resampled.}

\item{dfz}{a dataframe generated with \code{zenithSimulate}.}

\item{zenith}{the solar zenith angle that defines twilight.}

\item{dft}{a dataframe generated with \code{twilightSimulate}.}

\item{err}{a vector of adjustments (in minutes) to the twilight
times.}
}
\value{
\code{zenithSimulate} returns a data frame with
components
\item{\code{Date}}{times along the simulated track}
\item{\code{Lon}}{longitudes along the simulated track}
\item{\code{Lat}}{latitudes along the simulated track}
\item{\code{Zenith}}{zenith angles along the simulated track}
\code{twilightSimulate} returns a data frame of twilights with
components
\item{\code{Twilight}}{times of twilight}
\item{\code{Rise}}{is this a sunrise}
\item{\code{Lon}}{longitude at twilight}
\item{\code{Lat}}{latitude at twilight}
}
\description{
Simulate zenith angles, times and locations of twilight along a
specified track.
}
\details{
Given times, longitudes and latitudes that specify a template
track, \code{zenithSimulate} interpolates the template onto the
new times specified by \code{tm.out} and computes the solar zenith
angle at each point along the new track. Given a dataframe
generated by \code{zenithSimulate}, \code{twilightSimulate}
computes times and locations of sunrise and sunset based on the
simulated zenith angles. The \code{twilightPerturb} adds a given
vector of errors (in minutes) to the twilights in a dataframe
generated by \code{twilightSimulate}, in such a way that a
positive error causes sunrise to occur later and sunset earlier.
}

