% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_try.R
\name{driver_try}
\alias{driver_try}
\title{Automated driver creation and geckodriver selection. Using \code{\link{initialization_sgat}} instead is recommended.}
\usage{
driver_try()
}
\value{
Tries for driver connection
}
\description{
Automated driver creation and geckodriver selection. Using \code{\link{initialization_sgat}} instead is recommended.
}
\examples{
driver_try()
}
