% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_sgat.R
\name{initialization_sgat}
\alias{initialization_sgat}
\title{Iterate over driver_try until the driver is created}
\usage{
initialization_sgat()
}
\value{
"Driver connection"
}
\description{
Iterate over driver_try until the driver is created
}
\examples{
initialization_sgat()
}
