% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgat.R
\name{sgat}
\alias{sgat}
\title{Iteration over sgat_day on every day of the week and finally save the retrieved data as a csv file}
\usage{
sgat(
  lugar.a.buscar,
  tiempo.espera = 10,
  carpeta.guardado = "CSVs Concurrencias"
)
}
\arguments{
\item{carpeta.guardado}{}
}
\value{
dataframe with retrieved information for all weekdays
}
\description{
Iteration over sgat_day on every day of the week and finally save the retrieved data as a csv file
}
\examples{
sgat(lugar.a.buscar = "museo nacional de bellas artes, buenos aires, argentina")
}
