% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgat_day.R
\name{sgat_day}
\alias{sgat_day}
\title{Opens firefox, searches in Google lugar.a.buscar place on the dia.semana weekday. Then it retrieves the "Popular Times" and geolocation information. If the search finds said information, the function returns a data.frame, if not, NULL}
\usage{
sgat_day(lugar.a.buscar, dia.semana, tiempo.espera = 10)
}
\arguments{
\item{lugar.a.buscar}{What you want to search in Google}

\item{dia.semana}{Day of the week from when you want to retrieve the "Popular Times" information}

\item{tiempo.espera}{Time measure of how much time you think it's sufficient given your internet connection to load a Google search page and not finding the information,}
}
\value{
dataframe with retrieved information for single weekday
}
\description{
Opens firefox, searches in Google lugar.a.buscar place on the dia.semana weekday. Then it retrieves the "Popular Times" and geolocation information. If the search finds said information, the function returns a data.frame, if not, NULL
}
\examples{
sgat_day("museo nacional de bellas artes, buenos aires, argentina", "miercoles")
}
