% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ushuaia_restaurants.R
\name{ushuaia_restaurants}
\alias{ushuaia_restaurants}
\title{Retrieves a vector of Ushuaian's restaurants and bar's names and adresses}
\usage{
ushuaia_restaurants()
}
\value{
Vector of restaurants and bars in Ushuaia, Tierra del Fuego, Argentina
}
\description{
Retrieves a vector of Ushuaian's restaurants and bar's names and adresses
}
\examples{
ushiaia_restaurants()
}
